Adding automated tests
Here is an automated test for the parallel sorting. It tests 3 things: sorting a
file, merging 2 sorted files, and sorting numerically instead of
lexicographically. You can put the file anywhere, but it will look for the
parallel sort file in binaries/parallel_sort.py. This is stored in the variable
BINARY_UNDER_TEST if you want to change it.

Run the tests by doing:

python parallel_sort_tests.py

It runs under python2 + python3 and doesn't require installing anything else.